work.html: Code fix - Code alignment - Update
Fixed the syntax error by including </a> where GitHub logo in the footer and realigned the 3 lines statement into a single line.
Updated GitHub logo in the footer with the actual GitHub project url (https://github.com/ktvprasaad/profile-website).
Realignment of 3 lines of linkedin logo declaration in the footer into a single line.