Logo, portrait, button color and svg image values update
Updated logo height value to adjust to 100% according to the device screen size.
Included the class of portrait under media with the max-width 500px to ignore the float property.
Updated button background and text color as per the contrast rule.
Included svg's class to adjust to 100% according to the device screen size.